+++ crates/scheduler/src/call/crypto/cbc.rs
@@ -15,7 +15,7 @@
 use wasefire_applet_api::crypto::cbc::{self as api, Api};
 use wasefire_board_api::Api as Board;
 #[cfg(feature = "board-api-crypto-aes256-cbc")]
-use wasefire_board_api::applet::Memory as _;
+use wasefire_board_api::applet::{Memory as _, MemoryExt as _};
 #[cfg(feature = "board-api-crypto-aes256-cbc")]
 use wasefire_board_api::crypto::cbc::Api as _;
 #[cfg(feature = "board-api-crypto-aes256-cbc")]
@@ -48,8 +48,8 @@ fn encrypt<B: Board>(mut call: SchedulerCall<B, api::encrypt::Sig>) {
     let memory = call.applet().memory();
     let result = try bikeshed _ {
         ensure_support::<B>()?;
-        let key = memory.get(*key, 32)?.into();
-        let iv = memory.get(*iv, 16)?.into();
+        let key = memory.get_array::<32>(*key)?.into();
+        let iv = memory.get_array::<16>(*iv)?.into();
         let blocks = memory.get_mut(*ptr, *len)?;
         board::crypto::Aes256Cbc::<B>::encrypt(key, iv, blocks)?
     };
@@ -62,8 +62,8 @@ fn decrypt<B: Board>(mut call: SchedulerCall<B, api::decrypt::Sig>) {
     let memory = call.applet().memory();
     let result = try bikeshed _ {
         ensure_support::<B>()?;
-        let key = memory.get(*key, 32)?.into();
-        let iv = memory.get(*iv, 16)?.into();
+        let key = memory.get_array::<32>(*key)?.into();
+        let iv = memory.get_array::<16>(*iv)?.into();
         let blocks = memory.get_mut(*ptr, *len)?;
         board::crypto::Aes256Cbc::<B>::decrypt(key, iv, blocks)?
     };