+++ crates/scheduler/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.5.2-git
 
+### Minor
+
+- Support `crypto-common` bump from `wasefire-board-api`
+
 ### Patch
 
 - Disable `unused-features` lint
@@ -225,4 +229,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 11 -->
+<!-- Increment to skip CHANGELOG.md test: 12 -->